[FIX] 친구 도메인 API 저장 및 조회 방식 통일

+++ src/main/java/org/lxdproject/lxd/diary/dto/MemberDiarySummaryResponseDTO.java
@@ -13,7 +13,7 @@ public class MemberDiarySummaryResponseDTO {
     private String username;
     private String nickname;
     private Long diaryCount;
-    private Integer friendCount;
+    private Long friendCount;
     private RelationType relation;
 
     @JsonInclude(JsonInclude.Include.NON_NULL) // null 값도 반환하려면 이 부분 주석처리하면 됨
